Se quita el mixeo de pistas del repositorio
