+++ app/src/main/java/com/android/harmoniatpi/ui/screens/projectManagementScreen/ProjectManagementScreen.kt
@@ -26,7 +26,6 @@ import androidx.compose.foundation.text.appendInlineContent
 import androidx.compose.material.icons.Icons
 import androidx.compose.material.icons.automirrored.filled.ArrowBack
 import androidx.compose.material.icons.filled.Add
-import androidx.compose.material.icons.filled.Save
 import androidx.compose.material3.Button
 import androidx.compose.material3.Card
 import androidx.compose.material3.CardDefaults
@@ -114,19 +113,7 @@ fun ProjectManagementScreen(
                     navigationIconContentColor = MaterialTheme.colorScheme.secondary,
                     actionIconContentColor = MaterialTheme.colorScheme.secondary,
                     scrolledContainerColor = MaterialTheme.colorScheme.secondaryContainer
-                ),
-                actions = {
-                    IconButton(
-                        onClick = {
-                            viewModel.mixProjectTracks()
-                        }
-                    ) {
-                        Icon(
-                            imageVector = Icons.Default.Save,
-                            contentDescription = "Guardar",
-                        )
-                    }
-                }
+                )
             )
         }
     ) { padding ->
